Centralize package.json checks
